tags/tagrequirements.txt
numpy==2.0.2
psycopg2-binary
python-dotenv
loguru
tqdm
librosa